Updated search functions.
Now the place search is the primary function and almost all searches are conducted using this. The only exception being a search containing a postcode with no spacing, this is then checked and handled by the postcode search function. Additionally the " London" appending has been removed so as to not confuse the searches. London is now a filter for the results of the place search rather than United Kingdom as we're targeting London cyclers. 

This has been tested with a fair few places, the only issue is a generic street name can provide a different part of London with the same street name, the workaround is to also put in the borough or town the street is in.